backend/pyproject.toml
[project]
name = "readwise-backend"
version = "0.1.0"
description = "Readwise智能读书助手后端服务"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "pypdf2==3.0.1",
    "langchain>=0.2.0",
    "langgraph>=0.0.69",
    "python-dotenv==1.0.0",
    "requests==2.31.0",
    "pymongo==4.6.0",
    "celery==5.3.4",
    "redis==5.0.1",
    "pydantic>=2.8.0",
    "tqdm==4.66.1",
    "tenacity==8.2.3",
    "beautifulsoup4==4.12.2",
    "python-docx==1.1.0",
    "ebooklib==0.18",
    "python-multipart==0.0.6",
    "motor==3.3.2",
    "aiohttp==3.9.5",
    # RAG系统依赖
    "openai>=1.0.0",
    "qdrant-client>=1.7.0",
    "jieba>=0.42.1",
    "scikit-learn>=1.3.0",
    "numpy>=1.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]  # 包含当前目录的所有Python模块

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0"
]
